+++ include/seqan3/io/structure_file/format_vienna.hpp
@@ -39,9 +39,9 @@
 #include <seqan3/range/detail/misc.hpp>
 #include <seqan3/range/views/take.hpp>
 #include <seqan3/range/views/take_until.hpp>
-#include <seqan3/range/views/to.hpp>
 #include <seqan3/utility/char_operations/predicate.hpp>
 #include <seqan3/utility/detail/type_name_as_string.hpp>
+#include <seqan3/utility/views/to.hpp>
 
 namespace seqan3
 {
