+++ include/seqan3/utility/views/to.hpp
@@ -7,13 +7,24 @@
 
 /*!\file
  * \author Hannes Hauswedell <hannes.hauswedell AT fu-berlin.de>
- * \brief [DEPRECATED] Provides seqan3::views::to.
- * \deprecated This header will be removed in 3.1. Please use seqan3/utility/views/to.hpp instead.
+ * \brief Provides seqan3::views::to.
  */
 
 #pragma once
 
-#include <seqan3/utility/views/to.hpp>
+#include <range/v3/range/conversion.hpp>
 
-SEQAN3_DEPRECATED_HEADER(
-    "This header is deprecated and will be removed in SeqAn-3.1.0; Please #include <seqan3/utility/views/to.hpp> instead.")
+#include <seqan3/core/platform.hpp>
+#include <seqan3/std/ranges>
+
+namespace seqan3::views
+{
+
+/*!\brief A to view
+ * \ingroup views
+ * \details
+ * \noapi{This is currently range-v3's to implementation.}
+ */
+inline constexpr auto to = ::ranges::views::to;
+
+} // namespace seqan3::views
