+++ test/unit/alignment/pairwise/pairwise_alignment_collection_test_template.hpp
@@ -14,8 +14,8 @@
 
 #include <seqan3/alignment/pairwise/align_pairwise.hpp>
 #include <seqan3/alphabet/views/to_char.hpp>
-#include <seqan3/range/views/to.hpp>
 #include <seqan3/test/expect_range_eq.hpp>
+#include <seqan3/utility/views/to.hpp>
 #include <seqan3/utility/views/zip.hpp>
 
 #include "fixture/alignment_fixture.hpp"
