+++ test/unit/search/fm_index_cursor/bi_fm_index_cursor_test.cpp
@@ -5,15 +5,15 @@
 // shipped with this file and also available at: https://github.com/seqan/seqan3/blob/master/LICENSE.md
 // -----------------------------------------------------------------------------------------------------
 
+#include <seqan3/std/ranges>
 #include <string>
 #include <string_view>
 
 #include <seqan3/alphabet/nucleotide/dna4.hpp>
 #include <seqan3/alphabet/nucleotide/dna5.hpp>
 #include <seqan3/alphabet/views/char_to.hpp>
-#include <seqan3/range/views/to.hpp>
 #include <seqan3/search/fm_index/bi_fm_index.hpp>
-#include <seqan3/std/ranges>
+#include <seqan3/utility/views/to.hpp>
 
 #include "bi_fm_index_cursor_test_template.hpp"
