+++ test/unit/search/helper.hpp
@@ -8,13 +8,13 @@
 #pragma once
 
 #include <algorithm>
+#include <seqan3/std/iterator>
 #include <vector>
 
 #include <seqan3/core/detail/debug_stream_type.hpp>
-#include <seqan3/range/views/to.hpp>
 #include <seqan3/search/fm_index/bi_fm_index_cursor.hpp>
 #include <seqan3/search/fm_index/fm_index_cursor.hpp>
-#include <seqan3/std/iterator>
+#include <seqan3/utility/views/to.hpp>
 
 namespace seqan3
 {
